Cleanup Code for the doma-core module (#578)
* Remove unused instance fields

* Remove "unused" suppression

* Remove unnecessary `private`

* Remove unnecessary `static`, `public`, and `abstract`

* Replace with 'StandardCharsets.UTF_8'

* Remove redundant if statements

* Replace with 'String.join'

* Replace with enhanced 'for'

* Specify type parameter

* Remove boxing

* Remove unboxing

* Replace with <>

* Collapse 'catch' blocks

* Replace with 'contains()'

* Replace with lambda

* Replace lambda with method reference

* Replace with expression lambda

* Make `static`

* Avoid unpredictable BigDecimal constructor call

* Replace with 'Collections.singleList()'

* Suppress 'AssertBetweenInconvertibleTypes'

* Simplify assertion

* Remove explicit String creation

* Remove explicit Array creation

* Remove type arguments

* Suppress unnecessary 'continue' and 'return'

* Replace StringBuilder with String

* Remove redundant suppression

* Remove redundant suppression

* Suppress SameReturnValue

* Make final

* Suppress CStyleArrayDeclaration

* Remove unnecessary semicolon

* Print a message instead of an exception

* Replace with 'while'

* Replace with 'if'

* Inline variable

* Suppress BooleanMethodIsAlwaysInverted

* Remove redundant 'extends Object'

* Suppress OptionalUsedAsFieldOrParameterType

* Rename exception parameter name

* Replace with expression lambda

* Make 'static'

* Suppress DuplicateBranchesInSwitch

* Remove redundant initializer

* Suppress ImplicitArrayToString

* Suppress OptionalGetWithoutIsPresent

* Suppress ResultOfMethodCallIgnored

* Replace with 'while'

* Suppress UnusedAssignment

* Suppress SuspiciousInvocationHandlerImplementation

* Remove unnecessary test class

* Suppress EmptyMethod

* Convert octal literal to decimal literal

* Remove unnecessary condition

* Make 'final'

* Suppress SameParameterValue

* Replace with 'BigDecimal.valueOf'

* Suppress BigDecimalMethodWithoutRoundingCalled

* Suppress MismatchedQueryAndUpdateOfCollection

* Suppress CopyConstructorMissesField

* Suppress SameReturnValue

* Suppress NullableProblems

* Replace with 'null'

* Suppress ConstantConditions

* Remove redundant assignment

* Remove redundant if statement

* Split exception creation and exception throwing

* Use 'new HashSet()' instead of 'Collections.emptySet()' because the Set is mutable

* Remove unnecessary 'throws' declarations

* Remove unnecessary assertion

* Fix typo

* Avoid possible NullPointerException

* Suppress MagicConstant

* Suppress RedundantThrows

* Remove unnecessary 'throws' declarations

* Remove redundant suppression

* Extract common part

+++ doma-core/src/main/java/org/seasar/doma/wrapper/IntegerWrapper.java
@@ -22,7 +22,7 @@ public class IntegerWrapper extends AbstractWrapper<Integer> implements NumberWr
   public void increment() {
     Integer value = doGet();
     if (value != null) {
-      doSet(value.intValue() + 1);
+      doSet(value + 1);
     }
   }
 
@@ -30,7 +30,7 @@ public class IntegerWrapper extends AbstractWrapper<Integer> implements NumberWr
   public void decrement() {
     Integer value = doGet();
     if (value != null) {
-      doSet(value.intValue() - 1);
+      doSet(value - 1);
     }
   }
